backend/requirements.txt
sanic==23.12.1
aiomysql==0.2.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
sanic-cors==2.2.0
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
mysql-connector-python==8.3.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
setuptools==69.2.0
cryptography==42.0.5
pyotp==2.8.0
qrcode==7.4.2
pillow==10.0.0
openai==1.12.0 